ContentResolver moved to render package (from ui)
